Shut down binder instance when we are done with it
We should shut down binder instance when we are done, to
conserve resources and prevent ourselves from hitting usage
limits. Ideally, the token received from BinderHub can
be used to call the JupyterHub API to shut it down. However,
the token we get from BinderHub isn't a valid JupyterHub
token - just a valid Jupyter Notebook token. All JupyterHub
tokens are valid Jupyter Notebook tokens, but not all
Jupyter Notebook tokens are valid JupyterHub tokens! This
was a decision we made early on in the process so users
didn't need to have 'jupyterhub' installed inside their
containers.

We instead kill pid 1, which stops the container. JupyterHub
should garbage collect this shortly after. This is very
hacky, but works

Fixes #3