Refactor code into two classes
BinderBot class contained two bits of functionality:

1. Starting and stopping binders on a BinderHub
2. Interacting with the notebook API (start kernels, contentsmanager,
etc)

This commit splits them into two different classes. This has
the big advantage that we can run unit tests against a local instance
of jupyter notebooks, rather than starting an instance on mybinder.org
all the time. Tests are much faster, and everything works as
normal since it is a straight up split

We added more unit tests, but lost the CLI test. Most of its
functionality is covered by the other unit tests, but it would
be nice to have an integration test